feature(ui): Add SidePanel
